web-frontend/pyproject.toml
[project]
name = "web-frontend"
version = "0.3.0"
description = "Voice agent web frontend (LiveKit token mint + SuperTokens auth gate)"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "livekit-api",
    "supertokens-python",
]
